chore: harvest functionality with mainnet fork test

+++ src/notional/interfaces/ICToken.sol
@@ -2,9 +2,9 @@
 pragma solidity >0.8.8;
 
 interface ICToken {
-    function mint(uint assets) external returns (uint);
+    function mint(uint256 assets) external returns (uint256);
 
-    function redeem(uint redeemTokens) external returns (uint);
+    function redeem(uint256 redeemTokens) external returns (uint256);
 
-    function getCash() external view returns (uint);
+    function getCash() external view returns (uint256);
 }
